refactor(mcp): thin tools layer, service layering, consistent metadata
Tools-layer consistency cleanup. No tool names, input schemas, or output
shapes change (docs-fixtures diff is only cosmetic title casing).

- analytics/widget: move business logic + raw SQL out of the @McpTool
  classes into services; tool methods are now thin delegators
- extract the duplicated API-key minting and origin-allowlist checks into
  common/ helpers used by both analytics and widget
- rename the vendor channel-admin files/classes that carried no @McpTool
  (*.tools.ts/*AdminTools -> *-admin.service.ts/*AdminService) for Twilio,
  MessageBird, Vapi, Threll
- analytics inputs: named Zod consts + z.infer instead of inline schemas
  paired with hand-written arg types
- standardize empty-input on shared EmptyInput; always set both
  readOnlyHint and destructiveHint on feedback/system-alerts tools; fix
  system_alerts_* title casing
