+++ packages/backend-core/docs-fixtures/mcp-tools.json
@@ -7370,7 +7370,7 @@
   },
   {
     "name": "system_alerts_list",
-    "title": "System alerts: list",
+    "title": "System alerts: List",
     "description": "List operational alerts for the org. Defaults to open alerts only; pass includeResolved to see history.",
     "audiences": [
       "admin"
@@ -7396,7 +7396,7 @@
   },
   {
     "name": "system_alerts_get",
-    "title": "System alerts: get",
+    "title": "System alerts: Get",
     "description": "Read a single alert by id.",
     "audiences": [
       "admin"
@@ -7420,7 +7420,7 @@
   },
   {
     "name": "system_alerts_acknowledge",
-    "title": "System alerts: acknowledge",
+    "title": "System alerts: Acknowledge",
     "description": "Mark an alert as acknowledged. Does not resolve it; only signals that someone is on it.",
     "audiences": [
       "admin"
@@ -7444,7 +7444,7 @@
   },
   {
     "name": "system_alerts_resolve",
-    "title": "System alerts: resolve",
+    "title": "System alerts: Resolve",
     "description": "Manually resolve the open alert for the given source+subject. Writers normally self-resolve; use this when the underlying state can no longer be observed.",
     "audiences": [
       "admin"
